pyproject.toml
[tool.poetry]
name = "rag4p-jfall"
version = "0.1.0"
description = "Workshop for jfall"
authors = ["Jettro Coenradie <jettro.coenradie@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nltk = "^3.9.1"
openai = "^1.13"
pycryptodome = "^3.20.0"
requests = "^2.32.3"
configparser = "^6.0.0"
python-dotenv = "^1.0.0"
pandas = "^2.1.4"
scipy = "^1.11.4"
onnxruntime = "^1.18.0"
tokenizers = "^0.19.1"
unidecode = "^1.3.8"
weaviate-client = {version = "^4.0.0", allow-prereleases = false}
tiktoken = "^0.7.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
